Incorporate reviewer feedback
Remember to set the camera position for the MapActivity when a user does
not have a LatLon position.

Add animation to SelectorActivity and ChatActivity.

Create alumni in ChatActivity if they do not exist.

Expand Member.java to include a last_seen field that shows when the
person was last in the ChatActivity.

Check for null instances of Member.class in ProfileActivity.

Create a FirebaseHelper class so the table references all come from one
place.